refactor: create new select_builder class and fix bug where only mapped fields on parent object were translated
